backend_ai/requirements.txt
fastapi
uvicorn
google-generativeai>=0.7.2
python-dotenv
pydantic
groq
langgraph>=0.2.28
langchain-core>=0.3.0
langchain-groq>=0.2.0
openai>=1.40.0
pinecone-client>=5.0.0
slowapi>=0.1.9
PyJWT>=2.8.0
cryptography>=42.0.0
asyncpg>=0.29.0
huggingface_hub>=0.20.0
